Add get parking lot status API and get all vehicles in a parking lot API. Add flyway and migration scripts

+++ src/main/java/com/royeen/smartpark/business/usecase/GetParkingLotOccupancyCommand.java
@@ -1,5 +1,5 @@
 package com.royeen.smartpark.business.usecase;
 
 public interface GetParkingLotOccupancyCommand {
-    long execute(String lotId);
+    int execute(String lotId);
 }
